--- client/src/components/Wrapper/index.js
@@ -1,8 +1,0 @@
-import React from "react";
-import "./style.css";
-
-function Wrapper(props) {
-  return <main className="wrapper" {...props} />;
-}
-
-export default Wrapper;
