--- client/src/components/Wrapper/style.css
@@ -1,4 +1,0 @@
-.wrapper {
-  min-height: 100%;
-  padding-bottom: 100px;
-}
